Change repository name
This repository is still during its early development so I hope that
this change is not going to break anything for anyone.  I wanted to
change the name of this repository to make it more descriptive for
others at first glance.  The new name is also much more common through
Emacs configuration repositories available online.